\name{plot_boxkm}

\alias{plot_boxkm}

\title{Function for Visualization of Survival Distributions}

\description{
  Plot the cross-validated Kaplan-Meir estimates of survival distributions for the highest risk (inbox) versus
  lower-risk (outbox) groups of samples at each iteration of the peeling sequence (inner loop of our PRSP algorithm).
  
  Some plotting parameters are further defined in the function \code{plot.survfit} (R package \pkg{survival}).

}

\usage{
  plot_boxkm(peelobj,
             main = NULL, xlab = "Time", ylab = "Probability",
             precision, mark = 3, col = 1, lty = 1, lwd = 1, cex = 1,
             only.last = FALSE, nr = NULL, nc = NULL, ...)
}

\arguments{
  \item{peelobj}{
    Object of class "PRSP" as generated by the main function \code{\link[PRIMsrc]{sbh}}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. Defaults to \code{NULL}.
  }
  \item{xlab}{
    \code{Character} \code{vector}. X axis label. Defaults to "Time".
  }
  \item{ylab}{
    \code{Character} \code{vector}. Y axis label. Defaults to "Probability".
  }
  \item{precision}{
    Precision of cross-validated log-rank p-values of separation between two survival curves.
    Lower bounded by the value \eqn{\frac{1}{A}} of the number of replication \eqn{A}.
  }
  \item{mark}{
    \code{Integer} \code{vector} of mark parameters, which will be used to label the curves. Defaults to 3.
  }
  \item{col}{
    \code{Integer} \code{vector} of specifying colors for each curve. Defaults to 1.
  }
  \item{lty}{
    \code{Integer} \code{vector} of specifying line types for each curve. Defaults to 1.
  }
  \item{lwd}{
    \code{Numeric} \code{vector} of values for line widths. Defaults to 1.
  }
  \item{cex}{
    \code{Numeric} scalar specifying the size of the marks. Defults to 1.
  }
  \item{only.last}{
    \code{Logical} scalar defining whether only the last step of the peeling sequence should be plotted. Defaults to \code{FALSE}.
  }
  \item{nr}{
    \code{Integer} scalar of the number of rows in the plot. If \code{NULL}, defaults to 3.
  }
  \item{nc}{
    \code{Integer} scalar of the number of columns in the plot. If \code{NULL}, defaults to 4.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions, including \code{plot.survfit} (R package \pkg{survival}).
  }
}

\details{
  Step #0 always corresponds to the situation where the starting box covers the entire test-set data before peeling.
  Cross-validated LRT, LHR of inbox samples and log-rank p-values of separation are shown at the bottom of the plot
  with the corresponding peeling step. P-values are lower-bounded by the precision limit given by 1/#number of permutations (A).
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              (Submitted).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E. and J. S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\seealso{
    \itemize{
        \item \code{plot.survfit} (R package \pkg{survival})
    }
}

\examples{
#===================================================
# Loading the library and its dependencies
#===================================================
library("PRIMsrc")

\dontrun{
    #=================================================================================
    # Simulated dataset #1 (n=250, p=3)
    # Replicated Combined Cross-Validation (RCCV)
    # Peeling criterion = LRT
    # Optimization criterion = LRT
    #=================================================================================
    postscript(file=paste(getwd(), "/RCCV probability curves for model #1.ps", sep=""),
               horizontal = TRUE, width = 11.5, height = 8.5)
    plot_boxkm(peelobj = CVCOMBREP.synt1, 
               main = paste("RCCV probability curves for model #1", sep=""),
               xlab = "Time", ylab = "Probability",
               cex = 1, yscale = 1, precision = 1/CVCOMBREP.synt1$A)
    dev.off()
}

}
