\name{plot_boxtraj}

\alias{plot_boxtraj}

\title{Function for Visualization of Peeling Trajectory/Profiles}

\description{
  Plot the cross-validated peeling trajectories/profiles of covariates and other statistical quantities of interest
  at each iteration of the peeling sequence (inner loop of our PRSP algorithm). Plot up to three "PRSP"" objects.
}

\usage{
  plot_boxtraj(peelobj, peelobj2 = NULL, peelobj3 = NULL,
               main = NULL, xlab = "Box Mass", ylab = "Variable Range",
               col = 1, col2 = col, col3 = col, 
               lty = 1, lty2 = lty, lty3 = lty, 
               lwd = 1, lwd2 = lwd, lwd3 = lwd, 
               cex = 1, cex2 = cex, cex3 = cex,
               add = FALSE, add.legend = FALSE, text.legend = NULL, 
               nr = NULL, nc = NULL, ...)
}

\arguments{
  \item{peelobj}{
    Object of class "PRSP" as generated by the main function \code{\link[PrimSRC]{sbh}}.
  }
  \item{peelobj2}{
    Object of class "PRSP" as generated by the main function \code{\link[PrimSRC]{sbh}}.
    Defaults to \code{NULL}.
  }
  \item{peelobj3}{
    Object of class "PRSP" as generated by the main function \code{\link[PrimSRC]{sbh}}.
    Defaults to \code{NULL}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. Defaults to \code{NULL}.
  }
  \item{xlab}{
    \code{Character} \code{vector}. X axis label. Defaults to "Box Mass".
  }
  \item{ylab}{
    \code{Character} \code{vector}. Y axis label. Defaults to "Variable Range".
  }
  Plotting parameters for the three "PRSP"" objects:
  \item{col}{
    Line color of object peelobj. Defaults to 1.
  }
  \item{col2}{
    Line color of object peelobj2. Defaults to col.
  }
  \item{col3}{
    Line color of object peelobj3. Defaults to col.
  }
  \item{lty}{
    Line type of object peelobj. Defaults to 1.
  }
  \item{lty2}{
    Line type of object peelobj2. Defaults to lty.
  }
  \item{lty3}{
    Line type of object peelobj3. Defaults to lty.
  }
  \item{lwd}{
    Line width of object peelobj. Defaults to 1.
  }
  \item{lwd2}{
    Line width of object peelobj2. Defaults to lwd.
  }
  \item{lwd3}{
    Line width of object peelobj3. Defaults to lwd.
  }
  \item{cex}{
    Symbol expansion. Defaults to 1.
  }
  \item{cex2}{
    Symbol expansion2. Defaults to cex.
  }
  \item{cex3}{
    Symbol expansion3. Defaults to cex.
  }
  \item{add}{
    \code{Logical} scalar. Should this plot be added to an open graphics device?. Defaults to \code{FALSE}.
  }
  \item{add.legend}{
    \code{Logical} scalar. Should the legend be added to the current open graphics device?. Defaults to \code{FALSE}.
  }
  \item{text.legend}{
    \code{Character} \code{vector} of legend content. Defaults to \code{NULL}.
  }
  \item{nr}{
    \code{Integer} scalar of the number of rows in the plot. If \code{NULL}, defaults to 3.
  }
  \item{nc}{
    \code{Integer} scalar of the number of columns in the plot. If \code{NULL}, defaults to 3.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions.
  }
}

\details{
  The plot includes box descriptive statistics (such as support), survival endpoint statistics
  (such as Maximum Event-Free Time (MEFT), Minimum Event-Free Probability (MEVP), LHR, LRT)
  and prediction performance (such as CER).
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              (Submitted).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E. and J. S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jedazard@users.noreply.github.com}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jedazard@users.noreply.github.com}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\examples{
#===================================================
# Loading the library and its dependencies
#===================================================
library("PrimSRC")

#===================================================
# Loading of the Synthetic and Real datasets
# (see description of datasets)
#===================================================
data("Synthetic.1", "Real", package="PrimSRC")

\dontrun{
    #=================================================================================
    # Replicated Combined Cross-Validation (RCCV)
    # Simulated Synthetic.1 from model #1 (n=250, p=3)
    # Peeling criterion = LRT
    # Optimization criterion = LRT
    #=================================================================================
    if (CVCOMBREP.synt1$plot) {
        postscript(file=paste(getwd(), "/RCCV peeling traj. for model #1.ps", sep=""),
                   horizontal = FALSE, width = 8.5, height = 8.5)
        plot_boxtraj(peelobj = CVCOMBREP.synt1,
                     main = paste("RCCV peeling trajectories for model #1", sep=""),
                     col = 2:4, lty = c(1,1,1))
        dev.off()
    }
}

}
