\name{plot_surface}

\alias{plot_surface}

\title{Two-Dimensional Visualization for Model Selection/Validation}

\description{
  Function for plotting a perspective plot of the cross-validated tuning surface of a \code{CV} object. 
  It uses the user's choice of statistics among the Log Hazard Ratio (LHR), 
  Log-Rank Test (LRT) or Concordance Error Rate (CER) as a function of the optimal choice/number of 
  pre-selected variables and of the model tuning parameter, that is, and of number of peeling steps 
  of the peeling sequence (inner loop of our PRSP algorithm).
}

\usage{
  plot_surface(cvobj,
               main = NULL, 
               xlab = "Model Size (# variables)",
               ylab = "Peeling Steps",
               theta = 5, 
               phi = 10, 
               expand = 0.2, 
               col = "lightblue",
               add.line = FALSE, 
               col.line = "yellow",
               lty.line = 1, 
               lwd.line = 1, 
               pch.line = 20, 
               cex.line = 1,
               device = NULL, 
               file = "Surface Plot", 
               path=getwd(), 
               horizontal = FALSE, 
               width = 8.5, 
               height = 5.0, ...)
}

\arguments{
  \item{cvobj}{
    Object of class \code{CV} as generated by the main function \code{\link[PRIMsrc]{cv.sbh}}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. Defaults to \code{NULL}.
  }
  \item{xlab}{
    \code{Character} \code{vector}. X axis label. Defaults to "Model Size (# variables)".
  }
  \item{ylab}{
    \code{Character} \code{vector}. Y axis label. Defaults to "Peeling Steps".
  }
  \item{theta}{
    \code{Integer} scalar of angle defining the viewing direction of the perspective plot. 
    Defaults to 5.
    \code{theta} gives the azimuthal direction.
    See base function \code{persp} from R package \pkg{graphics}.
  }
  \item{phi}{
    \code{Integer} scalar of angle defining the viewing direction of the perspective plot. 
    Defaults to 10.
    \code{phi} gives the colatitude.
    See base function \code{persp} from R package \pkg{graphics}.
  }
  \item{expand}{
    \code{Integer} scalar of a expansion factor applied to the z coordinates. 
    Defaults to 0.2.
    Often used with 0 < expand < 1 to shrink the plotting box in the z direction. 
    See base function \code{persp} from R package \pkg{graphics}.
  }
  \item{col}{
    \code{Integer} scalar of color of the surface facets of the mean profiles. 
    Defaults to "lightblue".	
    Transparent colours are ignored.
    See base function \code{persp} from R package \pkg{graphics}.
  }
  \item{add.line}{
    \code{Logical} scalar. Shall the optimal path line be plotted on the surface? 
    Defaults to \code{FALSE}.
  }
  \item{col.line}{
    \code{Integer} scalar of line color of the optimal path line. 
    Defaults to "yellow".
  }
  \item{lty.line}{
    \code{Integer} scalar of line type of the optimal path line. 
    Defaults to 1.
  }
  \item{lwd.line}{
    \code{Integer} scalar of line width of the optimal path line. 
    Defaults to 1.
  }
  \item{pch.line}{
    \code{Integer} scalar of symbol number of the points on the optimal path line. 
    Defaults to 20.
  }
  \item{cex.line}{
    \code{Integer} scalar of symbol expansion of the points on the optimal path line. 
    Defaults to 1.
  }
  \item{device}{
    Graphic display device in \{NULL, "PS", "PDF"\}. Defaults to NULL (standard output screen).
    Currently implemented graphic display devices are "PS" (Postscript) or "PDF" (Portable Document Format).
  }
  \item{file}{
    File name for output graphic. Defaults to "Surface Plot".
  }
  \item{path}{
    Absolute path (without final (back)slash separator). Defaults to working directory path.
  }
  \item{horizontal}{
    \code{Logical} scalar. Orientation of the printed image. Defaults to \code{FALSE}, that is potrait orientation.
  }
  \item{width}{
    \code{Numeric} scalar. Width of the graphics region in inches. Defaults to 8.5.
  }
  \item{height}{
    \code{Numeric} scalar. Height of the graphics region in inches. Defaults to 5.0.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions.
  }
}

\details{
  Model tuning is done by applying the optimization criterion defined by the user's choice of specific statistic.
  The goal is to find the optimal values of the number of pre-selected variables and number of steps 
  by maximization of LHR or LRT, or minimization of CER.
  
  Currently, this done internally for visualization purposes, but it will ultimately offer the option to do be
  interactive with the end-user as well for parameter choosing/model selection.
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              Statistical Analysis and Data Mining (in press).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{R package PRIMsrc: Bump Hunting by Patient Rule Induction Method for Survival, Regression and Classification.}"
              In JSM Proceedings, Statistical Programmers and Analysts Section. Seattle, WA, USA.
              American Statistical Association IMS - JSM, (in press).
        \item Dazard J-E. and J.S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\examples{
#===================================================
# Loading the library and its dependencies
#===================================================
library("PRIMsrc")

#===================================================
# Simulated dataset #1 (n=250, p=3)
# Non Replicated Combined Cross-Validation (RCCV)
# Peeling criterion = LRT
# Optimization criterion = LRT
# Without parallelization
#===================================================
CVCOMB.CV <- cv.sbh(dataset = Synthetic.1,
                    B = 1, K = 5,
                    cvtype = "combined",
                    cvcriterion = "lrt",
                    conservative = "least",
                    arg = "beta=0.05,
                           alpha=0.05,
                           minn=5,
                           peelcriterion=\"lr\"",
                    fdr = NULL, thr = NULL,
                    parallel = FALSE, conf = NULL, seed = 123)

plot_surface(cvobj = CVCOMB.CV, 
             main = paste("Cross-validated tuning surface for model #1", sep=""),
             xlab = "Model Size (# variables)", ylab = "Peeling Steps",
             theta = 5, phi = 10, expand = 0.2, col = "lightblue", 
             add.line = TRUE,
             col.line = "yellow", lty.line = 1, lwd.line = 1, 
             pch.line = 20, cex.line = 1,
             device = NULL)
}
