\name{plot_trace}

\alias{plot_trace}

\title{Visualization of Covariates Traces}

\description{
  Function for plotting the cross-validated covariates traces of a \code{sbh} object.
  Plot the cross-validated modal trace curves of covariate importance and covariate usage of the 
  pre-selected covariates specified by user at each iteration of the peeling sequence 
  (inner loop of our PRSP or PRGSP algorithm). 
}

\usage{
  plot_trace(object,
             main = "Covariate Trace Plots",
             xlab = expression(paste("Box Support (", beta, ")", sep="")), 
             ylab = "Covariate Range (centered)",
             toplot = object$cvfit$cv.used,
             center = TRUE, 
             scale = FALSE, 
             col.cov, 
             lty.cov, 
             lwd.cov, 
             col = 1, 
             lty = 1, 
             lwd = 0.5, 
             cex = 0.5, 
             add.caption = FALSE, 
             text.caption = NULL,
             device = NULL, 
             file = "Covariate Trace Plots", 
             path = getwd(), 
             horizontal = FALSE, 
             width = 8.5, 
             height = 8.5, ...)
}

\arguments{
  \item{object}{
    Object of class \code{sbh} as generated by the main function \code{\link[PRIMsrc]{sbh}}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. 
    Defaults to "Covariate Trace Plots".
  }
  \item{xlab}{
    \code{Character} \code{vector}. X-axis label. 
    Defaults to "Box Support (beta)".
  }\code{NULL}
  \item{ylab}{
    \code{Character} \code{vector}. Y-axis label. 
    Defaults to "Covariate Range (centered)".
  }
  \item{toplot}{
    \code{Numeric} \code{vector}. Which of the pre-selected covariates to plot (in reference to the original index of covariates).
    Defaults to covariates used for peeling.
  }
  \item{center}{
    \code{Logical} scalar. Shall the data be centered?. 
    Defaults to \code{TRUE}.
  }
  \item{scale}{
    \code{Logical} scalar. Shall the data be scaled? 
    Defaults to \code{FALSE}.
  }
  \item{col.cov}{
    \code{Integer} \code{vector}. Line color for the covariate importance curve of each selected covariate. 
    Defaults to vector of colors of length the number of selected covariates. 
    The vector is reused cyclically if it is shorter than the number of selected covariates.
  }
  \item{lty.cov}{
    \code{Integer} \code{vector}. Line type for the covariate importance curve of each selected covariate. 
    Defaults to vector of 1's of length the number of selected covariates. 
    The vector is reused cyclically if it is shorter than the number of selected covariates.
  }
  \item{lwd.cov}{
    \code{Integer} \code{vector}. Line width for the covariate importance curve of each selected covariate. 
    Defaults to vector of 1's of length the number of selected covariates. 
    The vector is reused cyclically if it is shorter than the number of selected covariates.
  }
  \item{col}{
    \code{Integer} scalar. Line color for the covariate trace curve. 
    Defaults to 1.
  }
  \item{lty}{
    \code{Integer} scalar. Line type for the covariate trace curve. 
    Defaults to 1.
  }
  \item{lwd}{
    \code{Numeric} scalar. Line width for the covariate trace curve. 
    Defaults to 0.5.
  }
  \item{cex}{
    \code{Numeric} scalar. Symbol expansion used for titles, captions, and axis labels. 
    Defaults to 0.5.
  }
  \item{add.caption}{
    \code{Logical} scalar. Should the caption be plotted?. 
    Defaults to \code{FALSE}.
  }
  \item{text.caption}{
    \code{Character} \code{vector} of caption content. 
    Defaults to \code{NULL}.
  }
  \item{device}{
    Graphic display device in \{\code{NULL}, "PS", "PDF"\}. 
    Defaults to \code{NULL} (standard output screen).
    Currently implemented graphic display devices are "PS" (Postscript) or "PDF" (Portable Document Format).
  }
  \item{file}{
    File name for output graphic. Defaults to "Covariate Trace Plots".
  }
  \item{path}{
    Absolute path (without final (back)slash separator). 
    Defaults to working directory path.
  }
  \item{horizontal}{
    \code{Logical} scalar. Orientation of the printed image. 
    Defaults to \code{FALSE}, that is potrait orientation.
  }
  \item{width}{
    \code{Numeric} scalar. Width of the graphics region in inches. 
    Defaults to 8.5.
  }
  \item{height}{
    \code{Numeric} scalar. Height of the graphics region in inches. 
    Defaults to 8.5.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions.
  }
}

\details{
  The trace plots limit the display of traces to those only covariates that are used for peeling. 
  If centered, an horizontal black dotted line about 0 is added to the plot.
 
  Due to the variability induced by cross-validation and replication, it is possible that more than 
  one covariate be used for peeling at a given step. So, for simplicity of the trace plots, 
  only the modal or majority vote trace value (over the folds and replications of the cross-validation) is plotted. 
  
  The top plot shows the overlay of covariate importance curves for each covariate.
  The bottom plot shows the overlay of covariate usage curves for each covariate. It is a dicretized 
  view of covariate importance.
  
  Both point to the magnitude and order with which covariates are used along the peeling sequence.
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jean-eudes.dazard@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
        \item "J. Sunil Rao, Ph.D." \email{Rao@biostat.med.miami.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jean-eudes.dazard@case.edu}
}

\references{
    \itemize{
        \item Dazard J-E. and Rao J.S. (2021a).
              "\emph{Variable Selection Strategies for High-Dimensional Recursive Peeling-Based Survival Bump Hunting Models.}"
              (in prep).
        \item Dazard J-E. and Rao J.S. (2021b).
              "\emph{Group Bump Hunting by Recursive Peeling-Based Methods: Application to Survival/Risk Predictive Models.}"
              (in prep).
        \item Dazard J-E., Choe M., Pawitan Y., and Rao J.S. (2021c).
              "\emph{Identification and Characterization of Informative Prognostic Subgroups by Survival Bump Hunting.}"
              (in prep).
        \item Rao J.S., Huilin Y., and Dazard J-E. (2020).
              "\emph{Disparity Subtyping: Bringing Precision Medicine Closer to Disparity Science.}"
              Cancer Epidemiology Biomarkers & Prevention, 29(6 Suppl):C018.
        \item Yi C. and Huang J. (2017).
              "\emph{Semismooth Newton Coordinate Descent Algorithm for Elastic-Net Penalized Huber Loss Regression and Quantile Regression}."
              J. Comp Graph. Statistics, 26(3):547-557.
        \item Dazard J-E., Choe M., LeBlanc M., and Rao J.S. (2016).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              Statistical Analysis and Data Mining, 9(1):12-42.
        \item Dazard J-E., Choe M., LeBlanc M., and Rao J.S. (2015).
              "\emph{R package PRIMsrc: Bump Hunting by Patient Rule Induction Method for Survival, Regression and Classification.}"
              In JSM Proceedings, Statistical Programmers and Analysts Section. Seattle, WA, USA.
              American Statistical Association IMS - JSM, p. 650-664.
        \item Dazard J-E., Choe M., LeBlanc M., and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
    }
}

\section{Acknowledgments}{
    This work made use of the High Performance Computing Resource in the Core Facility for Advanced Research Computing at Case Western Reserve University. 
    This project was partially funded by the National Institutes of Health NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

