\name{predict.sbh}

\alias{predict.sbh}

\alias{predict}

\title{Predict Function}

\description{
  S3-method \code{predict} function to predict the box membership and box vertices 
  on an independent set, using a cross-validated \code{sbh} fitted object.
}

\usage{
  \method{predict}{sbh}(object, 
        newdata, 
        steps = 1:object$cvfit$cv.nsteps, 
        groups = NULL,
        na.action = na.omit, ...)
}

\arguments{
  \item{object}{
    Object of class \code{sbh} as generated by the main function \code{\link[PRIMsrc]{sbh}}.
  }
  \item{newdata}{
    A \code{numeric} \code{matrix} of same \eqn{p} input covariates as input data \code{object$X}.
    If \code{newdata} is a \code{numeric} \code{data.frame}, it is acceptable and will be coerced 
    to a \code{numeric} \code{matrix}. Discrete nominal covariates will be treated as ordinal variables. 
    \code{NA} missing values are not allowed. See details.
  }
  \item{steps}{
    \code{Integer} \code{vector}. Vector of peeling steps at which to predict the box memberships 
    and box vertices. Defaults to all the peeling steps of \code{sbh} object \code{object}.
  }
  \item{groups}{
    Same as \code{sbh} input: \code{character} or \code{numeric} \code{vector}, or \code{factor} of 
    group membership indicator variable, automatically converted into a \code{factor}. It is of length 
    the sample size, where entries take on group \code{levels}. Only two groups are allowed at this point.
    Defaults to \code{NULL}. See \code{sbh} for details.
  }
  \item{na.action}{
    A function to specify the action to be taken if NAs are found.
    The default action is na.omit, which leads to rejection of incomplete cases.
  }
  \item{\dots}{
    Further generic arguments passed to the predict function.
  }
}

\details{
    Only the used covariates of the final \code{sbh} object will be retained for the covariates of \code{newdata}. 
    So, the used covariates of \code{sbh} object must be equal or a subset of the 
    the covariates of \code{newdata}.
    
    Matrix \code{newdata} is usually an independent dataset drawn from the same population as the input 
    \code{X} dataset. It can also come from a split input dataset into two sets, usually for validation purposes.
}

\value{
    Object of \code{class} \code{sbh} (Patient Recursive Survival Peeling), where the \code{cvfit} list contains 
    the usual values of \code{sbh} object.    
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jean-eudes.dazard@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
        \item "J. Sunil Rao, Ph.D." \email{Rao@biostat.med.miami.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jean-eudes.dazard@case.edu}
}

\references{
    \itemize{
        \item Dazard J-E. and Rao J.S. (2021a).
              "\emph{Variable Selection Strategies for High-Dimensional Recursive Peeling-Based Survival Bump Hunting Models.}"
              (in prep).
        \item Dazard J-E. and Rao J.S. (2021b).
              "\emph{Group Bump Hunting by Recursive Peeling-Based Methods: Application to Survival/Risk Predictive Models.}"
              (in prep).
        \item Dazard J-E., Choe M., Pawitan Y., and Rao J.S. (2021c).
              "\emph{Identification and Characterization of Informative Prognostic Subgroups by Survival Bump Hunting.}"
              (in prep).
        \item Rao J.S., Huilin Y., and Dazard J-E. (2020).
              "\emph{Disparity Subtyping: Bringing Precision Medicine Closer to Disparity Science.}"
              Cancer Epidemiology Biomarkers & Prevention, 29(6 Suppl):C018.
        \item Yi C. and Huang J. (2017).
              "\emph{Semismooth Newton Coordinate Descent Algorithm for Elastic-Net Penalized Huber Loss Regression and Quantile Regression}."
              J. Comp Graph. Statistics, 26(3):547-557.
        \item Dazard J-E., Choe M., LeBlanc M., and Rao J.S. (2016).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              Statistical Analysis and Data Mining, 9(1):12-42.
        \item Dazard J-E., Choe M., LeBlanc M., and Rao J.S. (2015).
              "\emph{R package PRIMsrc: Bump Hunting by Patient Rule Induction Method for Survival, Regression and Classification.}"
              In JSM Proceedings, Statistical Programmers and Analysts Section. Seattle, WA, USA.
              American Statistical Association IMS - JSM, p. 650-664.
        \item Dazard J-E., Choe M., LeBlanc M., and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
    }
}

\section{Acknowledgments}{
    This work made use of the High Performance Computing Resource in the Core Facility for Advanced Research Computing at Case Western Reserve University. 
    This project was partially funded by the National Institutes of Health NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user predict function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}
