\name{updatecut}

\alias{updatecut}

\title{Internal Subroutine (never to be called by end-user)}

\description{
    Internal Subroutine of:
    \code{\link[PRIMsrc]{peel.box}}
    Update the splitpoints in the calling function.
}

\usage{
    updatecut(x, fract)
}

\arguments{
  \item{x}{
    \code{numeric} \code{matrix} of training covariates.
  }
  \item{fract}{
    \code{numeric} scalar quantile.
  }
}
