\name{plot.boxtrace}

\alias{plot.boxtrace.PRSP}
\alias{plot.boxtrace}

\title{Visualization of Covariates Traces}

\description{
  S3 generic function for plotting the cross-validated covariates traces of a \code{"PRSP"} object.
  Applies to the covariates used for peeling only. Plot the cross-validated modal trace curves 
  of variable importance and variable usage at each iteration of the peeling sequence 
  (inner loop of our PRSP algorithm). 
}

\usage{
  \method{plot.boxtrace}{PRSP}(x,
              main = NULL,
              center = FALSE, 
              scale = FALSE, 
              hline = NULL,
              col = 1, 
              lty = 1, 
              lwd = 1, 
              cex = 1, 
              add.legend = FALSE, 
              text.legend = NULL,
              device = NULL, 
              file = "Covariate Trace Plots", 
              path=getwd(), 
              horizontal = FALSE, 
              width = 8.5, 
              height = 8.5, ...)
}

\arguments{
  \item{x}{
    Object of class "PRSP" as generated by the main function \code{\link[PRIMsrc]{sbh}}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. Defaults to \code{NULL}.
  }
  \item{center}{
    \code{Logical} scalar. Shall the data be centered before?
  }
  \item{scale}{
    \code{Logical} scalar. Shall the data be scaled before?
  }
  \item{hline}{
    \code{Numeric} scalar of where the data is centred.
    If specified (i.e. not \code{NULL}), an horizontal line about this value is added to the plot.
  }
  \item{col}{
    Line color. Defaults to 1.
  }
  \item{lty}{
    Line type. Defaults to 1.
  }
  \item{lwd}{
    Line width. Defaults to 1.
  }
  \item{cex}{
    Symbol expansion. Defaults to 1.
  }
  \item{add.legend}{
    \code{Logical} scalar. Should the legend be added to the current open graphics device?. Defaults to \code{FALSE}.
  }
  \item{text.legend}{
    \code{Character} \code{vector} of legend content. Defaults to \code{NULL}.
  }
  \item{device}{
    Graphic display device in \{NULL, "PS", "PDF"\}. Defaults to NULL (standard output screen).
    Currently implemented graphic display devices are "PS" (Postscript) or "PDF" (Portable Document Format).
  }
  \item{file}{
    File name for output graphic. Defaults to "Covariate Trace Plots".
  }
  \item{path}{
    Absolute path (without final (back)slash separator). Defaults to working directory path.
  }
  \item{horizontal}{
    \code{Logical} scalar. Orientation of the printed image. Defaults to \code{FALSE}, that is potrait orientation.
  }
  \item{width}{
    \code{Numeric} scalar. Width of the graphics region in inches. Defaults to 8.5.
  }
  \item{height}{
    \code{Numeric} scalar. Height of the graphics region in inches. Defaults to 8.5.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions.
  }
}

\details{
  The trace plots limit the display of traces to those only covariates that are used for peeling.
  
  Due to the variability induced by cross-validation and replication, it is possible that more than one variable be used for peeling at a given step.
  So, for simplicity of the trace plots, only the modal or majority vote trace value (over the folds and replications of the cross-validation) is plotted. 
  
  The top plot shows the overlay of variable importance curves for each covariate.
  The bottom plot shows the overlay of variable usage curves for each covariate. It is a dicretized view of variable importance.
  
  Both point to the magnitude and order with which covariates are used along the peeling sequence.
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              (Submitted).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E. and J. S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\examples{
#===================================================
# Loading the library and its dependencies
#===================================================
library("PRIMsrc")

\dontrun{
    #=================================================================================
    # Simulated dataset #1 (n=250, p=3)
    # Replicated Combined Cross-Validation (RCCV)
    # Peeling criterion = LRT
    # Optimization criterion = LRT
    #=================================================================================
    plot.boxtrace(x = CVCOMBREP.synt1,
                  main = paste("RCCV trace plots for model #1", sep=""),
                  center = TRUE, scale = FALSE, 
                  hline = c(0,0,0), col = 2:4, lty = c(1,1,1),
                  device = NULL, file = "Covariate Trace Plots", path=getwd(),
                  horizontal = FALSE, width = 8.5, height = 8.5)
}

}

