\name{plot_boxtraj}

\alias{plot_boxtraj}

\title{Function for Visualization of Peeling Trajectory/Profiles}

\description{
  Plot the cross-validated peeling trajectories/profiles of covariates used for peeling 
  and other statistical quantities of interest at each iteration of the peeling sequence 
  (inner loop of our PRSP algorithm). 
}

\usage{
  plot_boxtraj(peelobj,
               main = NULL, xlab = "Box Mass", ylab = "Variable Range",
               col = 1, lty = 1, lwd = 1, cex = 1, 
               add.legend = FALSE, text.legend = NULL, 
               nr = NULL, nc = NULL, ...)
}

\arguments{
  \item{peelobj}{
    Object of class "PRSP" as generated by the main function \code{\link[PrimSRC]{sbh}}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. Defaults to \code{NULL}.
  }
  \item{xlab}{
    \code{Character} \code{vector}. X axis label. Defaults to "Box Mass".
  }
  \item{ylab}{
    \code{Character} \code{vector}. Y axis label. Defaults to "Variable Range".
  }
  Plotting parameters for the three "PRSP"" objects:
  \item{col}{
    Line color of object peelobj. Defaults to 1.
  }
  \item{lty}{
    Line type of object peelobj. Defaults to 1.
  }
  \item{lwd}{
    Line width of object peelobj. Defaults to 1.
  }
  \item{cex}{
    Symbol expansion. Defaults to 1.
  }
  \item{add.legend}{
    \code{Logical} scalar. Should the legend be added to the current open graphics device?. Defaults to \code{FALSE}.
  }
  \item{text.legend}{
    \code{Character} \code{vector} of legend content. Defaults to \code{NULL}.
  }
  \item{nr}{
    \code{Integer} scalar of the number of rows in the plot. If \code{NULL}, defaults to 3.
  }
  \item{nc}{
    \code{Integer} scalar of the number of columns in the plot. If \code{NULL}, defaults to 3.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions.
  }
}

\details{
  The plot limits the display of trajectories to those only covariates that are used for peeling.
  
  The plot includes box descriptive statistics (such as support), survival endpoint statistics
  (such as Maximum Event-Free Time (MEFT), Minimum Event-Free Probability (MEVP), LHR, LRT)
  and prediction performance (such as CER).
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              (Submitted).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E. and J. S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\examples{
#===================================================
# Loading the library and its dependencies
#===================================================
library("PrimSRC")

\dontrun{
    #=================================================================================
    # Simulated dataset #1 (n=250, p=3)
    # Replicated Combined Cross-Validation (RCCV)
    # Peeling criterion = LRT
    # Optimization criterion = LRT
    #=================================================================================
    if (CVCOMBREP.synt1$plot) {
        postscript(file=paste(getwd(), "/RCCV peeling traj. for model #1.ps", sep=""),
                   horizontal = FALSE, width = 8.5, height = 8.5)
        plot_boxtraj(peelobj = CVCOMBREP.synt1,
                     main = paste("RCCV peeling trajectories for model #1", sep=""),
                     col = 2:4, lty = c(1,1,1))
        dev.off()
    }
}

}
