\name{plot_scatter}

\alias{plot_scatter}

\title{Function for 2D Visualization of Data Scatter and Box Vertices}

\description{
  Plot in a plane the cross-validated Data Scatter and Box Vertices at a given peeling step of the peeling sequence
  (inner loop of our PRSP algorithm).
}

\usage{
  plot_scatter(peelobj,
               main = NULL, xlab = "", ylab = "",
               proj = c(1,2), splom = TRUE, boxes = TRUE,
               steps = peelobj$cvfit$cv.nsteps,
               add.legend = TRUE, pch = 16, cex = 0.7, col = 1, 
               box.col = 2, box.lty = 2, box.lwd = 1, ...)
}

\arguments{
  \item{peelobj}{
    Object of class "PRSP" as generated by the main function \code{\link[PrimSRC]{sbh}}.
  }
  \item{main}{
    \code{Character} \code{vector}. Main Title. Defaults to \code{NULL}.
  }
  \item{xlab}{
    \code{Character} \code{vector}. X axis label. Defaults to "".
  }
  \item{ylab}{
    \code{Character} \code{vector}. Y axis label. Defaults to "".
  } 
  \item{proj}{
    \code{Integer} \code{vector} of length two, specifying the dimensions of the projection plane. Defaults to c(1,2).
  }
  \item{splom}{
    \code{Logical} scalar. Shall the scatter plot of points inside the box(es) be plotted? Default to \code{TRUE}.
  }
  \item{boxes}{
    \code{Logical} scalar. Shall the box vertices be plotted or just the scatter of  points? Default to \code{TRUE}.
  }
  \item{steps}{
    \code{Integer} \code{vector}. Vector of peeling steps at which one wants to plot the box vertices. 
    Defaults to the last peeling step only.
  }
  \item{add.legend}{
    \code{Logical} scalar. Shall the legend of steps numbers be plotted? Defaults to \code{TRUE}.
  }
  \item{pch}{
    Symbol number for the scatter plot. Defaults to 16.
  }
  \item{cex}{
    Symbol expansion. Defaults to 0.7.
  }
  \item{col}{
    Symbol color. Defaults to 1.
  }
  \item{box.col}{
    Line color of the box vertices. Defaults to 2.
  }
  \item{box.lty}{
    Line type of box vertices. Defaults to 2.
  }
  \item{box.lwd}{
    Line width of box vertices. Defaults to 1.
  }
  \item{\dots}{
    Generic arguments passed to other plotting functions.
  }
}

\details{
  The scatterplot is drawn on a graphical device with geometrically equal scales on the \eqn{X} and \eqn{Y} axes.
}

\value{
  Invisible. None. Displays the plot(s) on the specified \code{device}.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              (Submitted).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E. and J. S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jxd101@case.edu}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\note{
  End-user plotting function.
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\examples{
#===================================================
# Loading the library and its dependencies
#===================================================
library("PrimSRC")

\dontrun{
    #=================================================================================
    # Simulated dataset #1 (n=250, p=3)
    # Replicated Combined Cross-Validation (RCCV)
    # Peeling criterion = LRT
    # Optimization criterion = LRT
    #=================================================================================
    postscript(file=paste(getwd(), "/Scatterplot for model #1.ps", sep=""),
               horizontal = FALSE, width = 5, height = 5)
    plot_scatter(peelobj=CVCOMBREP.synt1, 
                 main = paste("Scatter plot for model #1", sep=""),
                 proj = c(1,2), splom = TRUE, boxes = TRUE,
                 steps = CVCOMBREP.synt1$cvfit$cv.nsteps, 
                 add.legend = TRUE, cex = 0.5, 
                 box.col = "red", box.lty = 3, box.lwd = 1)
    dev.off()
}

}
