\name{PrimSRC-package}

\alias{PrimSRC-package}

\alias{PrimSRC}

\docType{package}

\title{Bump Hunting by Patient Rule Induction Method in Survival, Regression and Classification settings}

\description{Performs a Bump Hunting serach by Patient Rule Induction Method. The method generates decision rules delineating a region in the predictor space, where the response is larger than its average over the entire space. The region is shaped as a hyperdimensional box that is not necessarily contiguous. Assumptions are that the multivariate input variables can be discrete or continuous and the univariate response variable can be discrete (Classification), continuous (Regression) or a time-to-event, possibly censored (Survival). It is intended to handle high-dimensional multivariate datasets, where the number of variables far exceeds that of the samples (\eqn{p \gg n} paradigm).}

\details{
    The current version is the initial release (version 1.0.0) that only includes the case of low dimensional situations and a survival response.
    Ultimately, it will include all the features described above.
    New features will be added soon as they are available.

    \tabular{ll}{
      Package: \tab PrimSRC\cr
      Type: \tab Package\cr
      Version: \tab 1.0.0\cr
      Date: \tab 2015-01-20\cr
      License: \tab GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007\cr
    }

    The following describes all the end-user functions (not the internal subroutines) that are needed to run a complete procedure.
    The other internal subroutines are not to be called by the end-user at any time.
    For computational efficiency, some end-user functions offer a parallelization option that is done by passing
    a few parameters needed to configure a cluster. This is indicated by an asterisk (* = optionally involving cluster usage).
    The R functions are categorized as follows:

    \enumerate{

      \item{NEWS}
      \code{\link[PrimSRC]{PrimSRC.news}}
      \bold{Function to display the NEWS file of the \pkg{PrimSRC} package}\cr

      \item{END-USER SURVIVAL BUMP HUNTING FUNCTION} \cr
      \code{\link[PrimSRC]{sbh}} (*)
      \bold{Function for Cross-Validated Survival Bump Hunting} \cr
            Main and unique end-user function for fitting a cross-validated survival bump hunting model.
            Returns cross-validated "PRSP" object, as generated by our Patient Recursive Survival Peeling or PRSP algorithm
            at each iteration of the peeling sequence (inner loop of the PRSP algorithm).
            The object contains cross-validated estimates of all the decision-rules of covariates and other statistical quantities of interest.
            It also enables displaying results graphically of/for model tuning/selection, all peeling trajectories, variable traces,
            and survival distributions (see plotting functions below for more details).
            The function offers a few options such as the type of cross-validation desired (\eqn{K}-fold (replicated)-averaged or-combined),
            peeling and optimization critera for model fitting, tuning and selection and a few more parameters for the PRSP algorithm.
            The function takes advantage of the R package \pkg{parallel} for efficient parallel execution. It allows users to create a 
            cluster of workstations on a local and/or remote machine(s), enabling scaling-up with the number of CPU cores specified.
            Discrete (or nominal) variables should be made (or re-arranged into) ordinal variables.\cr

      \item{END-USER PLOTS FOR MODEL VALIDATION CHECKING AND GRAPHICAL VISUALIZATION OF RESULTS} \cr
      \code{\link[PrimSRC]{plot_profile}}
      \bold{Function for Model Validation Visualization and/or Checking} \cr
            Plot the cross-validated profiles of user's choice of statistics among the Log Hazard Ratio (LHR), Log-Rank Test (LRT),
            or Concordance Error Rate (CER) as a function of the model tuning parameter, that is, the optimal number of peeling steps
            of the peeling sequence (inner loop of our PRSP algorithm). Model validation is done by applying the optimization criterion 
            on the user's choice of specific statistic. The goal is to find the optimal value of the \eqn{K}-fold cross-validated number 
            of steps by maximization of LHR or LRT, or minimization of CER. Currently, this done internally for visualization purposes, 
            but it will ultimately offer the option to do be interactive with the end-user as well for parameter choosing/model selection. \cr
      \code{\link[PrimSRC]{plot_scatter}}
      \bold{Function for 2D Visualization of Data Scatter and Box Vertices} \cr
            Plot in a plane the cross-validated Data Scatter and Box Vertices at a given peeling step of the peeling sequence (inner loop of our PRSP algorithm).
            The scatterplot is drawn on a graphical device with geometrically equal scales on the \eqn{X} and \eqn{Y} axes. \cr
      \code{\link[PrimSRC]{plot_boxtraj}}
      \bold{Function for Visualization of Peeling Trajectory/Profiles} \cr
            Plot the cross-validated peeling trajectories/profiles of covariates and other statistical quantities of interest
            at each iteration of the peeling sequence (inner loop of our PRSP algorithm).
            The plot includes box descriptive statistics (such as support), survival endpoint statistics
            (such as Maximum Event-Free Time (MEFT), Minimum Event-Free Probability (MEVP), LHR, LRT)
            and prediction performance (such as CER). \cr
      \code{\link[PrimSRC]{plot_boxtrace}}
      \bold{Function for Visualization of Covariates Traces} \cr
            Plot the cross-validated trace curves of covariates variable importance and variable usage at each iteration of the peeling sequence
            (inner loop of our PRSP algorithm). The top plot shows the overlay of variable importance curves for each covariate
            The bottom plot shows the overlay of variable usage curves for each covariate. It is a dicretized view of variable importance.
            Both point to the magnitude and order in which covariates are used along the peeling sequence. \cr
      \code{\link[PrimSRC]{plot_boxkm}}
      \bold{Function for Visualization of Survival Distributions} \cr
            Plot the cross-validated Kaplan-Meir estimates of survival distributions for the highest risk (inbox) versus
            lower-risk (outbox) groups of samples at each iteration of the peeling sequence (inner loop of our PRSP algorithm).
            Step #0 always corresponds to the situation where the starting box covers the entire test-set data before peeling.
            Cross-validated LRT, LHR of inbox samples and log-rank p-values of separation are shown at the bottom of the plot
            with the corresponding peeling step. P-values are lower-bounded by the precision limit given by 1/#number of permutations (A). \cr

      \item{END-USER DATASETS} \cr
      \code{\link[PrimSRC]{Real}}
      \bold{Clinical Dataset} \cr
            Publicly available dataset from the Women's Interagency HIV cohort Study (WIHS).
            Inclusion criteria of the study were that women at enrolment were (i) alive, (ii) HIV-1 infected, and 
            (iii) free of clinical AIDS symptoms. Women were followed until the first of the following occurred: 
            (i) treatment initiation (HAART), (ii) AIDS diagnosis, (iii) death, or administrative censoring. 
            The studied outcomes were the competing risks "AIDS/Death (before HAART)" and "Treatment Initiation (HAART)".
            However, here, for simplification purposes, only the first of the two competing events (i.e. the time to AIDS/Death), 
            was used in this dataset example. Likewise, the entire study enrolled 1164 women, but only the complete cases were used 
            in this dataset example for simplification. Variables included history of Injection Drug Use ("IDU") at enrollment,
            African American ethnicity ("Race"), age ("Age"), and baseline CD4 count ("CD4"). The question in this dataset example 
            is whether it is possible to achieve a prognostication of patients for AIDS and HAART. \cr

      \code{\link[PrimSRC]{Synthetic.1}}, \code{\link[PrimSRC]{Synthetic.2}}, \code{\link[PrimSRC]{Synthetic.3}}, \code{\link[PrimSRC]{Synthetic.4}}
      \bold{Four Simulated Survival Models Datasets} \cr
            Modeling survival models #1-4 with censoring as a regression function of some informative predictors, depending on the model used. 
            In some models, this represents un-informative noisy variables, which are not part of the design matrix (models #2 and #3). 
            In one example, the signal is limited to a box-shaped region R of the predictor space (model #4). 
            Survival time was generated from an exponential model with with rate parameter \eqn{\lambda} (and mean \eqn{\frac{1}{\lambda}})
            according to a Cox-PH model with hazard exp(eta), where eta(.) is the regression function.
            Censoring indicator were generated from a uniform distribution on [0, 3].
            In these synthetic examples, all covariates are continuous, i.i.d. from a multivariate uniform distribution on [0, 1].
    }
    Known Bugs/Problems : None at this time.
}

\author{
    \itemize{
        \item "Jean-Eudes Dazard, Ph.D." \email{jedazard@users.noreply.github.com}
        \item "Michael Choe, M.D." \email{mjc206@case.edu}
        \item "Michael LeBlanc, Ph.D." \email{mleblanc@fhcrc.org}
        \item "Alberto Santana, MBA." \email{ahs4@case.edu}
    }
    Maintainer: "Jean-Eudes Dazard, Ph.D." \email{jedazard@users.noreply.github.com}

    Acknowledgments: This project was partially funded by the National Institutes of Health
    NIH - National Cancer Institute (R01-CA160593) to J-E. Dazard and J.S. Rao.
}

\references{
    \itemize{
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2015).
              "\emph{Cross-validation and Peeling Strategies for Survival Bump Hunting using Recursive Peeling Methods.}"
              (Submitted).
        \item Dazard J-E., Choe M., LeBlanc M. and Rao J.S. (2014).
              "\emph{Cross-Validation of Survival Bump Hunting by Recursive Peeling Methods.}"
              In JSM Proceedings, Survival Methods for Risk Estimation/Prediction Section. Boston, MA, USA.
              American Statistical Association IMS - JSM, p. 3366-3380.
        \item Dazard J-E. and J. S. Rao (2010).
              "\emph{Local Sparse Bump Hunting.}"
              J. Comp Graph. Statistics, 19(4):900-92.
    }
}

\keyword{Exploratory Survival/Risk Analysis}
\keyword{Survival/Risk Estimation & Prediction}
\keyword{Non-Parametric Method}
\keyword{Cross-Validation}
\keyword{Bump Hunting}
\keyword{Rule-Induction Method}

\seealso{
    \itemize{
        \item \code{makeCluster} (R package \pkg{parallel})
        \item \code{plot.survfit} (R package \pkg{survival})
        \item \code{wihs} (R package \pkg{randomForestSRC})
    }
}
